requirements.txt
streamlit
yt-dlp
youtube-transcript-api
groq
fpdf
python-dotenv
ffmpeg-python
